Avanzando hasta el punto 7

+++ MyBatis_Introduction_VideoRental/src/main/java/edu/eci/cvds/sampleprj/dao/ClienteDAO.java
@@ -5,5 +5,5 @@ import edu.eci.cvds.samples.entities.Cliente;
 public interface ClienteDAO {
 	public void save(Cliente it) throws PersistenceException;
 
-	public Cliente load(int id) throws PersistenceException;
+	public Cliente load(long id) throws PersistenceException;
 }
